Lights work. Spot light doesn't seem to have a cone that has 180 degree view.

+++ Graphics Engine/Graphics Engine/RenderMaterial.cpp
@@ -30,7 +30,7 @@ void RenderMaterial::Apply()
 
 	size_t count = m_ShaderResourceIDs.size();
 	for (size_t i = 0; i < count; i++) {
-		ShaderResourceManager::GetInstance()->ApplyShaderResource(m_ShaderResourceIDs[i], i);
+		ShaderResourceManager::GetInstance()->ApplyShaderResource(m_ShaderResourceIDs[i], (UINT)i);
 	}
 }
